improve comments and keep extensions im actually interested in
there are too many extensions and i shouldnt go crazy added everything
since i want to learn terminal development. only keep the bare minimum
that is useful in a jam, and stuff that only works on vscode. a lot of
it seems to not work with rust at the moment so left them as comments