fix joining zellij session and add project specfic template
i want to either attach or join a zellij session that might already exist, so a bash file
was added to facilitate that. i walso wanted to make project specific templates for zellij
so there are panes for everything i might need a click away. started doing it with alexandria